+++ app/models/__init__.py
@@ -11,213 +11,17 @@ from app.remote import usgs
 from .correlation import Correlation  # noqa
 from .user import User  # noqa
 from .region import Region  # noqa
+from .river import River  # noqa
+from .section import Section
 
-
-sections_regions = db.Table('sections_regions',
-    db.Column('section', db.Integer, db.ForeignKey('sections.id')),
-    db.Column('region', db.Integer, db.ForeignKey('regions.id'))
-)
 
 gages_regions = db.Table('gages_regions',
     db.Column('gage', db.Integer, db.ForeignKey('gages.id')),
     db.Column('region', db.Integer, db.ForeignKey('regions.id'))
 )
 
-rivers_regions = db.Table('rivers_regions',
-    db.Column('river', db.Integer, db.ForeignKey('rivers.id')),
-    db.Column('region', db.Integer, db.ForeignKey('regions.id'))
-)
 
 
-
-
-
-
-class River(db.Model):
-    """
-    River model. Rivers are rivers as we know them, \
-    they have a parent that they flow downstream into and have tributaries that flow into them.
-
-    Arguments:
-        id (int): Primary River id
-        name (string): Nice River name
-        slug (string): slug for url formatting
-        description (text): Long description that can contain HTML or Markdown within reason.
-        short_description (text): Short description for showing on other pages.
-        header_image (string): Header image to override default
-        parent_id (int): River.id for the river that this one flows into
-        parent: ``River`` object from parent_id
-        tributary: List of ``River`` objects for any River that has this one in it's parent_id.
-        regions: List of ``Region`` objects for River
-        sections: List of ``Section`` objects for River
-    """
-    __tablename__ = 'rivers'
-
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(80))
-    slug = db.Column(db.String(20), unique=True)
-    description = db.Column(db.Text)
-    short_description = db.Column(db.Text)
-    header_image = db.Column(db.String(80))
-
-    parent_id = db.Column(db.Integer, db.ForeignKey('rivers.id'))
-    parent = db.relationship('River', remote_side=id, backref='tributary')
-
-    regions = db.relationship('Region', secondary=rivers_regions,
-                              backref=db.backref('rivers', lazy='dynamic'))
-
-    #def __init__(self, name, slug, description,
-    #            short_description, header_image, parent):
-    #    self.name = name
-    #    self.slug = slug
-    #    self.description = description
-    #    self.short_description = short_description
-    #    self.header_image = header_image
-    #    self.parent = parent
-
-    def to_json(self):
-        """
-        Creates a JSON Object from River. Used where multiple rivers may be
-        listed at once.
-        """
-        json_river = {
-            'id': self.id,
-            'name': self.name,
-            'url': url_for('api.get_river', id=self.id, _external=True),
-        }
-        return json_river
-
-    def to_long_json(self):
-        """
-        Creates a JSON Object from River. Used where a single river is being
-        displayed.
-        """
-        if self.parent is not None:
-            json_river = {
-                'id': self.id,
-                'name': self.name,
-                'url': url_for('api.get_river', id=self.id, _external=True),
-                'sections': [section.to_json() for section in self.sections],
-                'downstream': self.parent.to_json(),
-                'tributaries': [river.to_json() for river in self.tributary],
-                'gages': [gage.to_json() for gage in self.gages]
-            }
-        else:
-            json_river = {
-                'id': self.id,
-                'name': self.name,
-                'url': url_for('api.get_river', id=self.id, _external=True),
-                'sections': [section.to_json() for section in self.sections],
-                'tributaries': [river.to_json() for river in self.tributary],
-                'gages': [gage.to_json() for gage in self.gages]
-            }
-        return json_river
-
-    def __repr__(self):
-        return '<River %r>' % self.name
-
-
-class Section(db.Model):
-    """
-    River section that is commonly paddled and possibly has a correlation developed for it.
-
-    Arguments:
-        id (int): Primary key for Section
-        name (string): Nice name for section
-        slug (string): slug used for url
-        river_id: Foreign River.id that this section is a segment of. \
-                    If a section spans multiple rivers, use the river that the gage is on.
-        river: ``River`` object
-        description (text): Long description for Section that can contain HTML or Markdown within reason.
-        short_description (text): Short description for section for showing on other pages.
-        access (string): Access restrictions or issues
-        location (string): Where is it?
-        putin (Point): PostGIS Point object for putin. Accepts WKT.
-        takeout (Point): PostGIS Point object for takeout. Accepts WKT.
-        path (Linestring): PostGIS Linestring object for river. Accepts WKT.
-        in_elevation (int): Elevation of the putin
-        out_elevation (int): Elevation of the takeout
-        header_image (string): Header image to override default
-        correlations: List of ``Correlation`` objects that connect this Section to Sensors. \
-                        Can do ``section.correlations[#].sensor`` and access the sensor attributes.
-    """
-    __tablename__ = 'sections'
-
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(80))
-    slug = db.Column(db.String(40), unique=True)
-
-    river_id = db.Column(db.Integer, db.ForeignKey('rivers.id'))
-    river = db.relationship('River', backref='sections')
-
-    description = db.Column(db.Text)
-    short_description = db.Column(db.Text)
-    access = db.Column(db.String)
-    location = db.Column(db.String)
-    putin = db.Column(Geometry('POINT'))
-    takeout = db.Column(Geometry('POINT'))
-    path = db.Column(Geometry('LINESTRING'))
-    in_elevation = db.Column(db.Integer)
-    out_elevation = db.Column(db.Integer)
-    header_image = db.Column(db.String(80))
-
-    regions = db.relationship('Region', secondary=sections_regions,
-                              backref=db.backref('sections', lazy='dynamic'))
-
-    def inlatlon(self):
-        """
-        Returns a shapely point for put-in.
-        section.inlatlon().y for latitude
-        section.inlatlon().y for longitude
-        """
-        try:
-            latlon_point = to_shape(self.putin)
-        except AssertionError:
-            return None
-        return latlon_point
-
-    def outlatlon(self):
-        """
-        Returns a shapely point for take-out.
-        section.outlatlon().y for latitude
-        section.outlatlon().x for longitude
-        """
-        try:
-            latlon_point = to_shape(self.takeout)
-        except AssertionError:
-            return None
-        return latlon_point
-
-    def to_json(self):
-        """
-        Creates a JSON Object from Section. Used where multiple sections may be listed at once.
-        """
-        json_section = {
-            'id' : self.id,
-            'name' : self.name,
-            'url': url_for('api.get_section', id=self.id, _external=True)
-        }
-        return json_section
-
-    def to_long_json(self):
-        """
-        Creates a JSON Object from Section. Used where a single section is displayed.
-        """
-        json_section = {
-            'id': self.id,
-            'name': self.name,
-            'url': url_for('api.get_section', id=self.id, _external=True),
-            'regions': [region.to_json() for region in self.regions],
-            'description': self.description,
-            'access': self.access,
-            'location': self.location,
-            'latitude': self.inlatlon().y,
-            'longitude': self.inlatlon().x,
-        }
-        return json_section
-
-    def __repr__(self):
-        return '<Section %r>' % self.name
 
 class Gage(db.Model):
     """
